Remove CommonErrorRegistry and register DittoJsonExceptions in GlobalErrorRegistry explicitly
