+++ signals/base/src/main/java/org/eclipse/ditto/signals/base/DittoJsonRegistry.java
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) 2017-2018 Bosch Software Innovations GmbH.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * which accompanies this distribution, and is available at
+ * https://www.eclipse.org/org/documents/epl-2.0/index.php
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+package org.eclipse.ditto.signals.base;
+
+import javax.annotation.concurrent.Immutable;
+
+@Immutable
+public final class DittoJsonRegistry {
+}
