+++ signals/base/src/main/java/org/eclipse/ditto/signals/base/GlobalErrorRegistry.java
@@ -20,8 +20,12 @@ import java.util.Map;
 import javax.annotation.concurrent.Immutable;
 
 import org.atteo.classindex.ClassIndex;
+import org.eclipse.ditto.json.JsonFieldSelectorInvalidException;
 import org.eclipse.ditto.json.JsonMissingFieldException;
 import org.eclipse.ditto.json.JsonObject;
+import org.eclipse.ditto.json.JsonParseException;
+import org.eclipse.ditto.json.JsonPointerInvalidException;
+import org.eclipse.ditto.model.base.exceptions.DittoJsonException;
 import org.eclipse.ditto.model.base.exceptions.DittoRuntimeException;
 import org.eclipse.ditto.model.base.headers.DittoHeaders;
 import org.eclipse.ditto.model.base.json.JsonParsableException;
@@ -39,7 +43,11 @@ public final class GlobalErrorRegistry extends AbstractErrorRegistry<DittoRuntim
     }
 
     private static Map<String, JsonParsable<DittoRuntimeException>> getParseRegistries() {
-        return JsonParsableExceptionRegistry.getParseRegistries();
+        final Map<String, JsonParsable<DittoRuntimeException>> parseRegistries =
+                new JsonParsableExceptionRegistry().getParseRegistries();
+        parseRegistries.putAll(new DittoJsonExceptionRegistry().getDittoJsonParseRegistries());
+
+        return parseRegistries;
     }
 
     @Override
@@ -54,12 +62,49 @@ public final class GlobalErrorRegistry extends AbstractErrorRegistry<DittoRuntim
         return instance;
     }
 
+
+    private static class DittoJsonExceptionRegistry {
+
+        private final Map<String, JsonParsable<DittoRuntimeException>> dittoJsonParseRegistries = new HashMap<>();
+
+        private DittoJsonExceptionRegistry() {
+
+            dittoJsonParseRegistries.put(JsonParseException.ERROR_CODE,
+                    (jsonObject, dittoHeaders) -> new DittoJsonException(
+                            JsonParseException.newBuilder().message(getMessage(jsonObject)).build(), dittoHeaders));
+
+            dittoJsonParseRegistries.put(JsonMissingFieldException.ERROR_CODE,
+                    (jsonObject, dittoHeaders) -> new DittoJsonException(
+                            JsonMissingFieldException.newBuilder().message(getMessage(jsonObject)).build(),
+                            dittoHeaders));
+
+            dittoJsonParseRegistries.put(JsonFieldSelectorInvalidException.ERROR_CODE,
+                    (jsonObject, dittoHeaders) -> new DittoJsonException(
+                            JsonFieldSelectorInvalidException.newBuilder().message(getMessage(jsonObject)).build(),
+                            dittoHeaders));
+
+            dittoJsonParseRegistries.put(JsonPointerInvalidException.ERROR_CODE,
+                    (jsonObject, dittoHeaders) -> new DittoJsonException(
+                            JsonPointerInvalidException.newBuilder().message(getMessage(jsonObject)).build(),
+                            dittoHeaders));
+        }
+
+        private static String getMessage(final JsonObject jsonObject) {
+            return jsonObject.getValue(DittoJsonException.JsonFields.MESSAGE)
+                    .orElseThrow(() -> JsonMissingFieldException.newBuilder()
+                            .fieldName(DittoRuntimeException.JsonFields.MESSAGE.getPointer().toString()).build());
+        }
+
+        private Map<String, JsonParsable<DittoRuntimeException>> getDittoJsonParseRegistries() {
+            return dittoJsonParseRegistries;
+        }
+
+    }
+
     private static class JsonParsableExceptionRegistry {
 
         private static final Class<?> JSON_OBJECT_PARAMETER = JsonObject.class;
         private static final Class<?> DITTO_HEADERS_PARAMETER = DittoHeaders.class;
-
-        private static final JsonParsableExceptionRegistry instance = new JsonParsableExceptionRegistry();
 
         private final Map<String, JsonParsable<DittoRuntimeException>> parseRegistries = new HashMap<>();
 
@@ -108,8 +153,8 @@ public final class GlobalErrorRegistry extends AbstractErrorRegistry<DittoRuntim
             return jsonParsableExceptions;
         }
 
-        static Map<String, JsonParsable<DittoRuntimeException>> getParseRegistries() {
-            return instance.parseRegistries;
+        private Map<String, JsonParsable<DittoRuntimeException>> getParseRegistries() {
+            return new HashMap<>(parseRegistries);
         }
     }
 }
